+++ simulations/vip-506/bscmainnet.ts
@@ -1,11 +1,11 @@
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import { vip501 } from "../../vips/vip-501/bsctestnet";
+import { vip506 } from "../../vips/vip-506/bscmainnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
-forking(51177545, async () => {
-  testVip("VIP-501", await vip501(), {
+forking(50348237, async () => {
+  testVip("VIP-506", await vip506(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [OMNICHAIN_PROPOSAL_SENDER_ABI], ["ExecuteRemoteProposal"], [1]);
     },
