+++ vips/vip-506/bsctestnet.ts
@@ -2,58 +2,58 @@ import { ethers } from "hardhat";
 import { LzChainId, ProposalType } from "src/types";
 import { makeProposal } from "src/utils";
 
-export const RESILIENT_ORACLE = "0xd2ce3fb018805ef92b8C5976cb31F84b4E295F94";
-export const CHAINLINK_ORACLE = "0x94c3A2d6B7B2c051aDa041282aec5B0752F8A1F2";
-export const REDSTONE_ORACLE = "0x0FC8001B2c9Ec90352A46093130e284de5889C86";
-export const BOUND_VALIDATOR = "0x1Cd5f336A1d28Dff445619CC63d3A0329B4d8a58";
-export const DEFAULT_PROXY_ADMIN = "0x567e4cc5e085d09f66f836fa8279f38b4e5866b9";
-export const RESILIENT_ORACLE_IMPLEMENTATION = "0x582d6d131e93D81676e82f032B2Dfa638F4E3275";
-export const CHAINLINK_ORACLE_IMPLEMENTATION = "0x36EFe8716fa2ff9f59D528d154D89054581866A5";
-export const REDSTONE_ORACLE_IMPLEMENTATION = "0xa3b4A56bf47a93459293CFA5E3D20c4f49C8643C";
-export const BOUND_VALIDATOR_IMPLEMENTATION = "0x955c01a8307618Ac3e5Fc08a7444f5cB6bD7d71e";
-export const weETH_ORACLE = "0xaB663D4a701229DFF407Eb4B45007921029072e9";
-export const weETH = "0xCd5fE23C85820F7B72D0926FC9b05b43E359b7ee";
-export const weETHsOracle = "0x47F7A7f3486b08A019E0c10Af969ADC4B6E415Cd";
-export const weETHs = "0x917ceE801a67f933F2e6b33fC0cD1ED2d5909D88";
-export const sFraxOracle = "0x1aDCE75BB3164bBf6060a4f44262df5414473110";
-export const sFrax = "0xA663B02CF0a4b149d2aD41910CB81e23e1c41c32";
-export const PTweETHOracle = "0xB89C0F93442C269271cB4e9Acd10E81D3fC237Ba";
-export const PTweETH = "0x6ee2b5E19ECBa773a352E5B21415Dc419A700d1d";
-export const PTUSDeOracle = "0x4CD93DcD2E11835D06a45F7eF9F7225C249Bb6Db";
-export const PTUSDe = "0x8A47b431A7D947c6a3ED6E42d501803615a97EAa";
-export const PTsUSDeOracle = "0x51B83bbbdCa078b2497C41c9f54616D1aDBEF86F";
-export const PTsUSDe = "0xE00bd3Df25fb187d6ABBB620b3dfd19839947b81";
-export const rsETH_Chainlink_Oracle = "0xc68A156b08C5C5C2e9c27B32A09977F3FA50FFE0";
-export const rsETH_Redstone_Oracle = "0x6AC694f2D118a35e1984AE590B916108F4f9337F";
-export const rsETH = "0xA1290d69c65A6Fe4DF752f95823fae25cB99e5A7";
-export const ezETH_Chainlink_Oracle = "0x84FAe9909Fa1F259CB23Fa14FCdd1a35A0FE7EB8";
-export const ezETH_Redstone_Oracle = "0xA6efeF98D9C4E9FF8193f80FbABF92AD92D50500";
-export const ezETH = "0xbf5495Efe5DB9ce00f80364C8B423567e58d2110";
-export const sUSDeOracle = "0xaE847E81ff6dD2bdFB1fD563ccB4f848c74D2B70";
-export const sUSDe = "0x9D39A5DE30e57443BfF2A8307A4256c8797A3497";
-export const sUSDSOracle = "0xEC3865a8a5DCb8C507781DA17A38b754E3d01C50";
-export const sUSDS = "0xa3931d71877C0E7a3148CB7Eb4463524FEc27fbD";
-export const yvUSDCOracle = "0xA938eFfc4f68262E17eAcfAFD41Ee68b0F8874D1";
-export const yvUSDC = "0xBe53A109B494E5c9f97b9Cd39Fe969BE68BF6204";
-export const yvUSDTOracle = "0xbC0289BFbF3F14af0829f10640a7B7E539910723";
-export const yvUSDT = "0x310B7Ea7475A0B449Cfd73bE81522F1B88eFAFaa";
-export const yvUSDSOracle = "0xE6eB6A0321CA8c18A3e4C31E36954467f72670EF";
-export const yvUSDS = "0x182863131F9a4630fF9E27830d945B1413e347E8";
-export const yvWETHOracle = "0x38b3643c1b5160591073cc4121Bd91A456F14Acd";
-export const yvWETH = "0xc56413869c6CDf96496f2b1eF801fEDBdFA7dDB0";
-export const eBTC = "0x657e8C867D8B37dCC18fA4Caead9C45EB088C642";
-export const eBTCOracle = "0x04d6096A6F089047C7af6E4644D18fB766B8d4cE";
-export const LBTC = "0x8236a87084f8B84306f72007F36F2618A5634494";
-export const LBTCOracle = "0x54B033D102db7DD734E0Ad649463E90fFA78D853";
-export const pufETH = "0xD9A442856C234a39a81a089C06451EBAa4306a72";
-export const pufETHOracle = "0x4fCbfE445396f31005b3Fd2F6DE2A986d6E2dCB5";
+export const RESILIENT_ORACLE = "0x8000eca36201dddf5805Aa4BeFD73d1EB4D23264";
+export const CHAINLINK_ORACLE = "0x102F0b714E5d321187A4b6E5993358448f7261cE";
+export const REDSTONE_ORACLE = "0x4e6269Ef406B4CEE6e67BA5B5197c2FfD15099AE";
+export const BOUND_VALIDATOR = "0x60c4Aa92eEb6884a76b309Dd8B3731ad514d6f9B";
+export const DEFAULT_PROXY_ADMIN = "0x01435866babd91311b1355cf3af488cca36db68e";
+export const RESILIENT_ORACLE_IMPLEMENTATION = "0xb4933AF59868986316Ed37fa865C829Eba2df0C7";
+export const CHAINLINK_ORACLE_IMPLEMENTATION = "0xDCB0CfA130496c749738Acbe2d6aA06C7C320f06";
+export const REDSTONE_ORACLE_IMPLEMENTATION = "0x6150997cd4B2f2366d1B0503F0DE020653b67BFe";
+export const BOUND_VALIDATOR_IMPLEMENTATION = "0xB634cd4F8b1CF2132E05381Eee0f994DF964869D";
+export const weETH_ORACLE = "0xf3ebD2A722c2039E6f66179Ad7F9f1462B14D8e0";
+export const weETH = "0x3b8b6E96e57f0d1cD366AaCf4CcC68413aF308D0";
+export const weETHsOracle = "0x660c6d8C5FDDC4F47C749E0f7e03634513f23e0e";
+export const weETHs = "0xE233527306c2fa1E159e251a2E5893334505A5E0";
+export const sFraxOracle = "0xac35DbBfACeDd301cB295DBeA57ddA7A032F0E41";
+export const sFrax = "0xd85FfECdB4287587BC53c1934D548bF7480F11C4";
+export const PTweETHOracle = "0x5CDE9fec66D89B931fB7a5DB8cFf2cDb642f4e7d";
+export const PTweETH = "0x56107201d3e4b7Db92dEa0Edb9e0454346AEb8B5";
+export const PTUSDeOracle = "0x16D54113De89ACE580918D15653e9C0d1DE05C35";
+export const PTUSDe = "0x74671106a04496199994787B6BcB064d08afbCCf";
+export const PTsUSDeOracle = "0x1bB3faB3813267d5b6c2abE5A284C621350544aD";
+export const PTsUSDe = "0x3EBa2Aa29eC2498c2124523634324d4ce89c8579";
+export const rsETH_Chainlink_Oracle = "0x3a6f2c02ec48dbEE4Ca406d701DCA2CC9d919EaD";
+export const rsETH_Redstone_Oracle = "0xA1eA3cB0FeA73a6c53aB07CcC703Dc039D8EAFb4";
+export const rsETH = "0xfA0614E5C803E15070d31f7C38d2d430EBe68E47";
+export const ezETH_Chainlink_Oracle = "0x50196dfad5030ED54190F75e5F9d88600A4CA0B4";
+export const ezETH_Redstone_Oracle = "0x987010fD82FDCe099174aC605B88E1cc35019ef4";
+export const ezETH = "0xB8eb706b85Ae7355c9FE4371a499F50f3484809c";
+export const sUSDeOracle = "0x93e19584359C6c5844f1f7E1621983418b5A892F";
+export const sUSDe = "0xA3A3e5ecEA56940a4Ae32d0927bfd8821DdA848A";
+export const sUSDSOracle = "0xb844Fc5cE6e9c0D0cB6E2336B0F86A3e4146DbfB";
+export const sUSDS = "0xE9E34fd81982438E96Bd945f5810F910e35F0165";
+export const yvUSDCOracle = "0x23c151c861c38544E5D09D7aBEDf2a4Cce1AfDa2";
+export const yvUSDC = "0x9fE6052B9534F134171F567dAC9c9B22556c1DDb";
+export const yvUSDTOracle = "0xcB99fcc1De353E69F970B499F945B41B3E35B404";
+export const yvUSDT = "0x5cBA66C5415E56CC0Ace55148ffC63f61327478B";
+export const yvUSDSOracle = "0x729d134355774224f7F65a85EF76b0EF08B31400";
+export const yvUSDS = "0xC6A0e98B8D9E9F1160E9cE1f2E0172F41FB06BC2";
+export const yvWETHOracle = "0x5fA77768b9b5927A1Ed4bc2C562AE375Cd1887A2";
+export const yvWETH = "0x99AD7ecf9b1C5aC2A11BB00D7D8a7C54fCd41517";
+export const eBTC = "0xd48392CCf3fe028023D0783E570DFc71996859d7";
+export const eBTCOracle = "0x51Fa73926Cd9456A94221b70428aa94F009a5b81";
+export const LBTC = "0x37798CaB3Adde2F4064afBc1C7F9bbBc6A826375";
+export const LBTCOracle = "0x75C573Bf23FbD0bff9B764c037C9543a1690A23b";
+export const pufETH = "0x6D9f78b57AEeB0543a3c3B32Cc038bFB14a4bA68";
+export const pufETHOracle = "0xbD2272b9f426dF6D18468fe5117fCFd547D6882b";
 
-export const ACM = "0x230058da2D23eb8836EC5DB7037ef7250c56E25E";
-export const NORMAL_TIMELOCK = "0xd969E79406c35E80750aAae061D402Aab9325714";
-export const CRITICAL_TIMELOCK = "0xeB9b85342c34F65af734C7bd4a149c86c472bC00";
-export const FASTTRACK_TIMELOCK = "0x8764F50616B62a99A997876C2DEAaa04554C5B2E";
+export const ACM = "0xbf705C00578d43B6147ab4eaE04DBBEd1ccCdc96";
+export const NORMAL_TIMELOCK = "0xc332F7D8D5eA72cf760ED0E1c0485c8891C6E0cF";
+export const CRITICAL_TIMELOCK = "0xA24A7A65b8968a749841988Bd7d05F6a94329fDe";
+export const FASTTRACK_TIMELOCK = "0x7F043F43Adb392072a3Ba0cC9c96e894C6f7e182";
 
-export const vip501 = () => {
+export const vip506 = () => {
   const meta = {
     version: "v2",
     title: "",
@@ -69,79 +69,79 @@ export const vip501 = () => {
         target: DEFAULT_PROXY_ADMIN,
         signature: "upgrade(address,address)",
         params: [RESILIENT_ORACLE, RESILIENT_ORACLE_IMPLEMENTATION],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: DEFAULT_PROXY_ADMIN,
         signature: "upgrade(address,address)",
         params: [CHAINLINK_ORACLE, CHAINLINK_ORACLE_IMPLEMENTATION],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: DEFAULT_PROXY_ADMIN,
         signature: "upgrade(address,address)",
         params: [REDSTONE_ORACLE, REDSTONE_ORACLE_IMPLEMENTATION],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: DEFAULT_PROXY_ADMIN,
         signature: "upgrade(address,address)",
         params: [BOUND_VALIDATOR, BOUND_VALIDATOR_IMPLEMENTATION],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setSnapshot(uint256,uint256)", NORMAL_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setGrowthRate(uint256,uint256)", NORMAL_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setSnapshotGap(uint256)", NORMAL_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setSnapshot(uint256,uint256)", CRITICAL_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setGrowthRate(uint256,uint256)", CRITICAL_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setSnapshotGap(uint256)", CRITICAL_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setSnapshot(uint256,uint256)", FASTTRACK_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setGrowthRate(uint256,uint256)", FASTTRACK_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setSnapshotGap(uint256)", FASTTRACK_TIMELOCK],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
       {
         target: RESILIENT_ORACLE,
@@ -242,7 +242,7 @@ export const vip501 = () => {
             ],
           ],
         ],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
     ],
     meta,
@@ -250,4 +250,4 @@ export const vip501 = () => {
   );
 };
 
-export default vip501;
+export default vip506;
